+++ src/routes/+layout.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import Header from './Header.svelte';
+	import Navbar from '../lib/components/Navbar.svelte';
 	import '../app.css';
 	interface Props {
 		children?: import('svelte').Snippet;
@@ -8,7 +8,7 @@
 	let { children }: Props = $props();
 </script>
 
-<Header />
+<Navbar />
 
 <main>
 	{@render children?.()}
